# digdag-docs/src/_static/swagger.yaml
...
swagger: "2.0"
info:
  description: "Digdag server API"
  version: "0.9.41-SNAPSHOT"
  title: "Digdag"
...
paths:
...
  /api/projects/{id}/workflow:
    get:
      tags:
      - "Project"
      operationId: "getWorkflow"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      - name: "revision"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinition"
          headers: {}
  /api/projects/{id}/workflows:
    get:
      tags:
      - "Project"
      summary: "List workflows of a project with filters"
      description: ""
      operationId: "getWorkflows"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "project id"
        required: true
        type: "integer"
        format: "int32"
      - name: "revision"
        in: "query"
        description: "use a given revision of the project instead of the latest revision"
        required: true
        type: "string"
      - name: "name"
        in: "query"
        description: "exact matching filter on workflow name"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinitionCollection"
  /api/projects/{id}/workflows/{name}:
    get:
      tags:
      - "Project"
      operationId: "getWorkflowByName"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - name: "name"
        in: "path"
        required: true
        type: "string"
      - name: "revision"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinition"
          headers: {}
...
  /api/workflow:
    get:
      tags:
      - "Workflow"
      operationId: "getWorkflowDefinition"
      produces:
      - "application/json"
      parameters:
      - name: "project"
        in: "query"
        required: false
        type: "string"
      - name: "revision"
        in: "query"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinition"
          headers: {}
  /api/workflows:
    get:
      tags:
      - "Workflow"
      summary: "List workflows"
      description: ""
      operationId: "getWorkflowDefinitions"
      produces:
      - "application/json"
      parameters:
      - name: "last_id"
        in: "query"
        description: "list workflows whose id is grater than this id for pagination"
        required: false
        type: "integer"
        format: "int64"
      - name: "count"
        in: "query"
        description: "number of workflows to return"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinitionCollection"
  /api/workflows/{id}:
    get:
      tags:
      - "Workflow"
      summary: "Get a workflow"
      description: ""
      operationId: "getWorkflowDefinition"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "workflow id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowDefinition"
  /api/workflows/{id}/truncated_session_time:
    get:
      tags:
      - "Workflow"
      summary: "Get truncated local time based on the time zone of a workflow"
      description: ""
      operationId: "getWorkflowDefinition"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "workflow id"
        required: true
        type: "integer"
        format: "int64"
      - name: "session_time"
        in: "query"
        description: "session time to be truncated"
        required: true
        type: "string"
      - name: "mode"
        in: "query"
        description: "truncation mode - second, minute, hour, day, schedule, or next_schedule"
        required: false
        type: "string"
        enum:
        - "SECOND"
        - "MINUTE"
        - "HOUR"
        - "DAY"
        - "SCHEDULE"
        - "NEXT_SCHEDULE"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestWorkflowSessionTime"
definitions:
...
  Config:
    type: "object"
    properties:
      empty:
        type: "boolean"
        default: false
      factory:
        $ref: "#/definitions/ConfigFactory"
      keys:
        type: "array"
        items:
          type: "string"
...
  IdAndName:
    type: "object"
    properties:
      name:
        type: "string"
      id:
        $ref: "#/definitions/Id"
  ZoneOffset:
    type: "object"
    properties:
      totalSeconds:
        type: "integer"
        format: "int32"
      id:
        type: "string"
      rules:
        $ref: "#/definitions/ZoneRules"
...
  RestWorkflowDefinitionCollection:
    type: "object"
    properties:
      workflows:
        type: "array"
        items:
          $ref: "#/definitions/RestWorkflowDefinition"
...
  RestWorkflowDefinition:
    type: "object"
    properties:
      project:
        $ref: "#/definitions/IdAndName"
      name:
        type: "string"
      id:
        $ref: "#/definitions/Id"
      config:
        $ref: "#/definitions/Config"
      revision:
        type: "string"
      timezone:
        readOnly: true
        $ref: "#/definitions/ZoneId"
  RestWorkflowSessionTime:
    type: "object"
    properties:
      project:
        $ref: "#/definitions/IdAndName"
      revision:
        type: "string"
      sessionTime:
        $ref: "#/definitions/OffsetDateTime"
      timezone:
        readOnly: true
        $ref: "#/definitions/ZoneId"
...
  ZoneOffsetTransition:
    type: "object"
    properties:
      offsetBefore:
        $ref: "#/definitions/ZoneOffset"
      offsetAfter:
        $ref: "#/definitions/ZoneOffset"
      overlap:
        type: "boolean"
        default: false
      instant:
        type: "integer"
        format: "int64"
      duration:
        $ref: "#/definitions/Duration"
      gap:
        type: "boolean"
        default: false
      dateTimeBefore:
        type: "string"
        format: "date-time"
      dateTimeAfter:
        type: "string"
        format: "date-time"
...
  ZoneOffsetTransitionRule:
    type: "object"
    properties:
      month:
        type: "string"
        enum:
        - "JANUARY"
        - "FEBRUARY"
        - "MARCH"
        - "APRIL"
        - "MAY"
        - "JUNE"
        - "JULY"
        - "AUGUST"
        - "SEPTEMBER"
        - "OCTOBER"
        - "NOVEMBER"
        - "DECEMBER"
      timeDefinition:
        type: "string"
        enum:
        - "UTC"
        - "WALL"
        - "STANDARD"
      standardOffset:
        $ref: "#/definitions/ZoneOffset"
      offsetBefore:
        $ref: "#/definitions/ZoneOffset"
      offsetAfter:
        $ref: "#/definitions/ZoneOffset"
      dayOfMonthIndicator:
        type: "integer"
        format: "int32"
      localTime:
        $ref: "#/definitions/LocalTime"
      midnightEndOfDay:
        type: "boolean"
        default: false
      dayOfWeek:
        type: "string"
        enum:
        - "MONDAY"
        - "TUESDAY"
        - "WEDNESDAY"
        - "THURSDAY"
        - "FRIDAY"
        - "SATURDAY"
        - "SUNDAY"
...
  LocalTime:
    type: "object"
    properties:
      hour:
        type: "integer"
        format: "int32"
      minute:
        type: "integer"
        format: "int32"
      second:
        type: "integer"
        format: "int32"
      nano:
        type: "integer"
        format: "int32"
  ZoneId:
    type: "object"
    properties:
      rules:
        $ref: "#/definitions/ZoneRules"
      id:
        type: "string"
...
  TemporalUnit:
    type: "object"
    properties:
      duration:
        $ref: "#/definitions/Duration"
      durationEstimated:
        type: "boolean"
        default: false
      dateBased:
        type: "boolean"
        default: false
      timeBased:
        type: "boolean"
        default: false
...
  Duration:
    type: "object"
    properties:
      seconds:
        type: "integer"
        format: "int64"
      nano:
        type: "integer"
        format: "int32"
      zero:
        type: "boolean"
        default: false
      negative:
        type: "boolean"
        default: false
      units:
        type: "array"
        items:
          $ref: "#/definitions/TemporalUnit"
...
  ConfigFactory:
    type: "object"
...
  OffsetDateTime:
    type: "object"
    properties:
      offset:
        $ref: "#/definitions/ZoneOffset"
      nano:
        type: "integer"
        format: "int32"
      hour:
        type: "integer"
        format: "int32"
      minute:
        type: "integer"
        format: "int32"
      dayOfYear:
        type: "integer"
        format: "int32"
      dayOfWeek:
        type: "string"
        enum:
        - "MONDAY"
        - "TUESDAY"
        - "WEDNESDAY"
        - "THURSDAY"
        - "FRIDAY"
        - "SATURDAY"
        - "SUNDAY"
      month:
        type: "string"
        enum:
        - "JANUARY"
        - "FEBRUARY"
        - "MARCH"
        - "APRIL"
        - "MAY"
        - "JUNE"
        - "JULY"
        - "AUGUST"
        - "SEPTEMBER"
        - "OCTOBER"
        - "NOVEMBER"
        - "DECEMBER"
      dayOfMonth:
        type: "integer"
        format: "int32"
      year:
        type: "integer"
        format: "int32"
      second:
        type: "integer"
        format: "int32"
      monthValue:
        type: "integer"
        format: "int32"
...
  ZoneRules:
    type: "object"
    properties:
      fixedOffset:
        type: "boolean"
        default: false
      transitions:
        type: "array"
        items:
          $ref: "#/definitions/ZoneOffsetTransition"
      transitionRules:
        type: "array"
        items:
          $ref: "#/definitions/ZoneOffsetTransitionRule"
...
  Id:
    type: "object"
...